# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: NeonVibe API
  description: "Servidor de música personal self-hosted. Los endpoints bajo /api/v1\
    \ requieren un JWT de acceso (header Authorization: Bearer <token>), salvo /auth/**\
    \ y /public/**."
  version: v1
servers:
- url: https://neonvibe.fepdev.app
  description: Producción
security:
- bearer: []
paths:
...
  /api/v1/favorites:
    get:
      tags:
      - favorite-controller
      operationId: list_2
      parameters:
      - name: entityType
        in: query
        required: false
        schema:
          type: string
          enum:
          - TRACK
          - ALBUM
          - ARTIST
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FavoriteResponse"
    post:
      tags:
      - favorite-controller
      operationId: create_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FavoriteRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/FavoriteResponse"
  /api/v1/auth/refresh:
    post:
      tags:
      - auth-controller
      operationId: refresh
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /api/v1/auth/google:
    post:
      tags:
      - auth-controller
      operationId: googleLogin
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoogleTokenRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /api/v1/artists/{id}/cover:
    get:
      tags:
      - cover-controller
      operationId: artistCover
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
                format: byte
    post:
      tags:
      - cover-controller
      operationId: uploadArtistCover
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required:
              - file
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ArtistResponse"
  /api/v1/albums/{id}/cover:
    get:
      tags:
      - cover-controller
      operationId: albumCover
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
                format: byte
    post:
      tags:
      - cover-controller
      operationId: uploadAlbumCover
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required:
              - file
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AlbumResponse"
...
  /api/v1/tracks:
    get:
      tags:
      - track-controller
      operationId: list_3
      parameters:
      - name: q
        in: query
        required: false
        schema:
          type: string
      - name: artist
        in: query
        required: false
        schema:
          type: string
      - name: album
        in: query
        required: false
        schema:
          type: string
      - name: genre
        in: query
        required: false
        schema:
          type: string
      - name: year
        in: query
        required: false
        schema:
          type: integer
          format: int32
      - name: pageable
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageTrackResponse"
  /api/v1/tracks/{id}:
    get:
      tags:
      - track-controller
      operationId: getById_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TrackResponse"
...
  /api/v1/tracks/{id}/cover:
    get:
      tags:
      - cover-controller
      operationId: trackCover
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
                format: byte
...
  /api/v1/artists:
    get:
      tags:
      - artist-controller
      operationId: list_4
      parameters:
      - name: q
        in: query
        required: false
        schema:
          type: string
      - name: pageable
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageArtistResponse"
  /api/v1/artists/{id}:
    get:
      tags:
      - artist-controller
      operationId: getById_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ArtistResponse"
  /api/v1/artists/{id}/tracks:
    get:
      tags:
      - artist-controller
      operationId: getTracks
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrackResponse"
  /api/v1/artists/{id}/albums:
    get:
      tags:
      - artist-controller
      operationId: getAlbums
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AlbumResponse"
  /api/v1/albums:
    get:
      tags:
      - album-controller
      operationId: list_5
      parameters:
      - name: q
        in: query
        required: false
        schema:
          type: string
      - name: artist
        in: query
        required: false
        schema:
          type: string
      - name: pageable
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageAlbumResponse"
  /api/v1/albums/{id}:
    get:
      tags:
      - album-controller
      operationId: getById_3
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AlbumResponse"
  /api/v1/albums/{id}/tracks:
    get:
      tags:
      - album-controller
      operationId: getTracks_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: pageable
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrackResponse"
...
  /api/v1/favorites/{id}:
    delete:
      tags:
      - favorite-controller
      operationId: delete_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
components:
  schemas:
...
    FavoriteRequest:
      type: object
      properties:
        entityType:
          type: string
          enum:
          - TRACK
          - ALBUM
          - ARTIST
        entityId:
          type: integer
          format: int64
      required:
      - entityId
      - entityType
    FavoriteResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        entityType:
          type: string
          enum:
          - TRACK
          - ALBUM
          - ARTIST
        entityId:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
    RefreshTokenRequest:
      type: object
      properties:
        refreshToken:
          type: string
          minLength: 1
      required:
      - refreshToken
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        tokenType:
          type: string
        expiresIn:
          type: integer
          format: int64
    GoogleTokenRequest:
      type: object
      properties:
        idToken:
          type: string
          minLength: 1
      required:
      - idToken
    ArtistResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        createdAt:
          type: string
          format: date-time
    AlbumResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        artist:
          type: string
        year:
          type: integer
          format: int32
        genre:
          type: string
        coverArtPath:
          type: string
        createdAt:
          type: string
          format: date-time
        trackCount:
          type: integer
          format: int64
    Pageable:
      type: object
      properties:
        page:
          type: integer
          format: int32
          minimum: 0
        size:
          type: integer
          format: int32
          minimum: 1
        sort:
          type: array
          items:
            type: string
    PageTrackResponse:
      type: object
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: "#/components/schemas/TrackResponse"
        number:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/SortObject"
        pageable:
          $ref: "#/components/schemas/PageableObject"
        numberOfElements:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
        empty:
          type: boolean
    PageableObject:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        sort:
          $ref: "#/components/schemas/SortObject"
        paged:
          type: boolean
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        unpaged:
          type: boolean
    SortObject:
      type: object
      properties:
        empty:
          type: boolean
        sorted:
          type: boolean
        unsorted:
          type: boolean
    TrackResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        filePath:
          type: string
        title:
          type: string
        artist:
          type: string
        album:
          type: string
        albumArtist:
          type: string
        year:
          type: integer
          format: int32
        genre:
          type: string
        trackNumber:
          type: integer
          format: int32
        discNumber:
          type: integer
          format: int32
        durationSeconds:
          type: integer
          format: int32
        bitrate:
          type: integer
          format: int32
        format:
          type: string
        mimeType:
          type: string
        hasLyrics:
          type: boolean
        coverArtPath:
          type: string
        isAvailable:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    PageArtistResponse:
      type: object
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: "#/components/schemas/ArtistResponse"
        number:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/SortObject"
        pageable:
          $ref: "#/components/schemas/PageableObject"
        numberOfElements:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
        empty:
          type: boolean
    PageAlbumResponse:
      type: object
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: "#/components/schemas/AlbumResponse"
        number:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/SortObject"
        pageable:
          $ref: "#/components/schemas/PageableObject"
        numberOfElements:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
        empty:
          type: boolean
...
  securitySchemes:
    bearer:
      type: http
      name: bearer
      scheme: bearer
      bearerFormat: JWT